Refactored builder into separate folder
