+++ src/brevitas_examples/common/quantizer_builder/describe_demo.py
@@ -8,19 +8,19 @@ test_input_quantizer_builder.py) and print, for each, the injector attributes,
 dependency kinds, and -- for ``@value`` functions -- the args they require and
 what they resolve to.
 
-Run with: python src/brevitas_examples/common/describe_quantizer_demo.py
+Run with: python src/brevitas_examples/common/quantizer_builder/describe_demo.py
 """
 from brevitas.core.stats.stats_op import AbsMax
 from brevitas.inject.enum import QuantType
 from brevitas.inject.enum import RestrictValueType
 from brevitas.inject.enum import ScalingImplType
 from brevitas.inject.enum import ScalingPerOutputType
-from brevitas_examples.common.input_quantizer_builder import build_input_quantizer
-from brevitas_examples.common.quantizer_builder import FloatFormat
-from brevitas_examples.common.quantizer_builder import ParamMethod
-from brevitas_examples.common.quantizer_builder import QuantParamType
-from brevitas_examples.common.quantizer_builder import ScaleType
-from brevitas_examples.common.weight_quantizer_builder import build_weight_quantizer
+from brevitas_examples.common.quantizer_builder.input import build_input_quantizer
+from brevitas_examples.common.quantizer_builder.mixins import FloatFormat
+from brevitas_examples.common.quantizer_builder.mixins import ParamMethod
+from brevitas_examples.common.quantizer_builder.mixins import QuantParamType
+from brevitas_examples.common.quantizer_builder.mixins import ScaleType
+from brevitas_examples.common.quantizer_builder.weight import build_weight_quantizer
 
 BIT_WIDTH = 8
 GROUP_SIZE = 8
