+++ src/brevitas_examples/common/quantizer_builder/weight.py
@@ -14,21 +14,21 @@ from brevitas.inject.enum import QuantType
 from brevitas.inject.enum import RestrictValueType
 from brevitas.inject.enum import ScalingImplType
 from brevitas.inject.enum import ScalingPerOutputType
-from brevitas_examples.common.quant_builder_components import CommonComponent
-from brevitas_examples.common.quant_builder_components import FormatComponent
-from brevitas_examples.common.quant_builder_components import ScaleComponent
-from brevitas_examples.common.quant_builder_components import ScaleParamMethodComponent
-from brevitas_examples.common.quant_builder_components import ScaleRestrictComponent
-from brevitas_examples.common.quant_builder_components import WeightIntQuantComponent
-from brevitas_examples.common.quant_builder_components import WeightSolverComponent
-from brevitas_examples.common.quant_builder_components import ZeroPointComponent
-from brevitas_examples.common.quant_builder_components import ZeroPointParamMethodComponent
-from brevitas_examples.common.quant_builder_core import Component
-from brevitas_examples.common.quant_builder_core import config_from_flat_args
-from brevitas_examples.common.quantizer_builder import FloatFormat
-from brevitas_examples.common.quantizer_builder import ParamMethod
-from brevitas_examples.common.quantizer_builder import QuantParamType
-from brevitas_examples.common.quantizer_builder_v2 import QuantizerBuilder
+from brevitas_examples.common.quantizer_builder.builder import QuantizerBuilder
+from brevitas_examples.common.quantizer_builder.components import CommonComponent
+from brevitas_examples.common.quantizer_builder.components import FormatComponent
+from brevitas_examples.common.quantizer_builder.components import ScaleComponent
+from brevitas_examples.common.quantizer_builder.components import ScaleParamMethodComponent
+from brevitas_examples.common.quantizer_builder.components import ScaleRestrictComponent
+from brevitas_examples.common.quantizer_builder.components import WeightIntQuantComponent
+from brevitas_examples.common.quantizer_builder.components import WeightSolverComponent
+from brevitas_examples.common.quantizer_builder.components import ZeroPointComponent
+from brevitas_examples.common.quantizer_builder.components import ZeroPointParamMethodComponent
+from brevitas_examples.common.quantizer_builder.core import Component
+from brevitas_examples.common.quantizer_builder.core import config_from_flat_args
+from brevitas_examples.common.quantizer_builder.mixins import FloatFormat
+from brevitas_examples.common.quantizer_builder.mixins import ParamMethod
+from brevitas_examples.common.quantizer_builder.mixins import QuantParamType
 
 
 class WeightQuantizerBuilder(QuantizerBuilder):
